weddingwebsite/requirements.txt
Django == 2.0.5
gunicorn == 19.8.1
whitenoise == 4.1.2
django-lockdown == 1.6.0
django-material == 1.2.2
Pillow == 5.00
django-imagekit == 4.0.2
